requirements.txt
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
distro==1.9.0
docstring_parser==0.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
soupsieve==2.8.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
